+++ packages/esense_flutter/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.6.2
+* Added podspec file to iOS
+
+## 0.6.1
+* fixed bug in sampling rate configuration on Android
+* small improvement to API doc and example app
+
+## 0.6.0
+* moved permission handling to app level
+* improved on iOS implementation
+* improvement to exampel app
+* improvement to README
+
+## 0.5.0
+* upgrade to Android embedding v2
+
+## 0.4.3
+* fix of error in setSamplingRate method on iOS
+
+## 0.4.0
+* upgrade to null-safety
+
 ## 0.3.0
 * singleton changed to recommended dart style using `ESenseManager()`.
 * fixed bug in reconnection.
@@ -8,7 +30,7 @@
 * updated Swift implementation to version 5
 
 ## 0.2.0+2
-* update to readme on iOS 
+* update to readme on iOS
 * added bluetooth permission to `Info.plist`
 
 ## 0.2.0+1
@@ -29,4 +51,4 @@
 * update of readme documentation on API.
 
 ## 0.1.0
-* Initial release supporting 1st released version of the Android eSense library, dated April 2019. 
+* Initial release supporting 1st released version of the Android eSense library, dated April 2019.